[ci] same for updateHistory
