+++ apps/scan/src/main/scala/org/lfdecentralizedtrust/splice/scan/store/ScanKeyValueProvider.scala
@@ -8,31 +8,26 @@ import com.digitalasset.canton.logging.{NamedLoggerFactory, NamedLogging}
 import io.circe.Codec
 import io.circe.generic.semiauto.deriveCodec
 import org.lfdecentralizedtrust.splice.store.{KeyValueStore, TimestampWithMigrationId}
-import cats.data.OptionT
 import cats.implicits.toBifunctorOps
-import com.digitalasset.canton.tracing.TraceContext
-import org.lfdecentralizedtrust.splice.scan.store.ScanKeyValueProvider.*
 import org.lfdecentralizedtrust.splice.scan.store.bulk.UpdatesSegment
-
-import scala.concurrent.{ExecutionContext, Future}
 
 class ScanKeyValueProvider(val store: KeyValueStore, val loggerFactory: NamedLoggerFactory)
     extends NamedLogging {
 
-  private val latestUpdatesSegmentInBulkStorageKey = "latest_updates_segment_in_bulk_storage"
-
-  final def setLatestUpdatesSegmentInBulkStorage(
-      segment: UpdatesSegment
-  )(implicit tc: TraceContext): Future[Unit] = store.setValue(
-    latestUpdatesSegmentInBulkStorageKey,
-    segment,
-  )
-
-  final def getLatestUpdatesSegmentInBulkStorage()(implicit
-      tc: TraceContext,
-      ec: ExecutionContext,
-  ): OptionT[Future, UpdatesSegment] =
-    store.readValueAndLogOnDecodingFailure(latestUpdatesSegmentInBulkStorageKey)
+//  private val latestUpdatesSegmentInBulkStorageKey = "latest_updates_segment_in_bulk_storage"
+//
+//  final def setLatestUpdatesSegmentInBulkStorage(
+//      segment: UpdatesSegment
+//  )(implicit tc: TraceContext): Future[Unit] = store.setValue(
+//    latestUpdatesSegmentInBulkStorageKey,
+//    segment,
+//  )
+//
+//  final def getLatestUpdatesSegmentInBulkStorage()(implicit
+//      tc: TraceContext,
+//      ec: ExecutionContext,
+//  ): OptionT[Future, UpdatesSegment] =
+//    store.readValueAndLogOnDecodingFailure(latestUpdatesSegmentInBulkStorageKey)
 }
 
 object ScanKeyValueProvider {
