+++ apps/scan/src/main/scala/org/lfdecentralizedtrust/splice/scan/store/bulk/AcsSnapshotBulkStorage.scala
@@ -3,6 +3,7 @@
 
 package org.lfdecentralizedtrust.splice.scan.store.bulk
 
+import com.daml.metrics.api.MetricHandle
 import com.digitalasset.canton.data.CantonTimestamp
 import com.digitalasset.canton.logging.{NamedLoggerFactory, NamedLogging}
 import com.digitalasset.canton.time.Clock
@@ -23,6 +24,8 @@ import org.lfdecentralizedtrust.splice.store.{TimestampWithMigrationId, UpdateHi
 import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.duration.*
 
+/** An abstract class for pipelines that process ACS snapshots for bulk storage.
+  */
 abstract class AcsSnapshotBulkStorage(
     appConfig: BulkStorageConfig,
     acsSnapshotStore: AcsSnapshotStore,
@@ -33,21 +36,41 @@ abstract class AcsSnapshotBulkStorage(
     extends NamedLogging
     with Spanning {
 
+  protected val description: String
+
+  /** The key in the key-value store where the timestamp of the latest processed snapshot is stored. This is
+    * used to resume processing from the correct point in case of restarts.
+    */
   protected val kvStoreKey: String
 
+  /** A metric that should be updated with the timestamp of the latest snapshot that was processed.
+    */
+  protected val processedTimestampMetric: MetricHandle.Gauge[CantonTimestamp]
+
+  /** This method should return the timestamp of the next snapshot available, after `last`, if any.
+    * The pipeline will  poll this method until it returns a new snapshot, and then process that snapshot
+    * before polling for the next one. It is ok for this method to return a snapshot that should not actually
+    * be processed, as the pipeline will call `shouldProcessSnapshotAt` to check if the snapshot should be
+    * processed or skipped (but will remember that it was skipped and update `last` for the future calls accordingly).
+    */
   protected def getNextSnapshotTimestampAfter(
       last: TimestampWithMigrationId
   )(implicit tc: TraceContext): Future[Option[TimestampWithMigrationId]]
 
+  /** This method should return true if the snapshot at the given timestamp should be processed,
+    * or false if it should be skipped. This is used to skip snapshots that are not relevant for bulk storage
+    * (e.g. because they are in the DB, but are more frequent then the frequency we need for bulk storage).
+    */
   protected def shouldProcessSnapshotAt(ts: TimestampWithMigrationId)(implicit
       tc: TraceContext
   ): Boolean
 
+  /** This method should return the main Flow that processes the snapshot at the given timestamp.
+    * It must emit back the same timestamp as its output once processing is complete.
+    */
   protected def processSnapshotAt(ts: TimestampWithMigrationId)(implicit
       tc: TraceContext
   ): Flow[TimestampWithMigrationId, TimestampWithMigrationId, NotUsed]
-
-  protected def updateMetric(ts: TimestampWithMigrationId): Unit
 
   protected[bulk] def readLatestProcessedSnapshotTimestamp(implicit
       tc: TraceContext
@@ -125,11 +148,11 @@ abstract class AcsSnapshotBulkStorage(
         .flatMapConcat {
           case Some(start: TimestampWithMigrationId) =>
             logger.info(
-              s"Latest dumped snapshot was from migration ${start.migrationId}, timestamp ${start.timestamp}"
+              s"Latest processed snapshot was from migration ${start.migrationId}, timestamp ${start.timestamp}"
             )
             getAcsSnapshotTimestampsAfter(start)
           case None =>
-            logger.info("No dumped snapshots yet, starting from genesis")
+            logger.info("No processed snapshots yet, starting from genesis")
             getAcsSnapshotTimestampsAfter(TimestampWithMigrationId(CantonTimestamp.MinValue, 0))
         }
         .filter { shouldProcessSnapshotAt }
@@ -139,7 +162,7 @@ abstract class AcsSnapshotBulkStorage(
             .via(processSnapshotAt(ts))
         )
         .mapAsync(1) { ts =>
-          updateMetric(ts)
+          processedTimestampMetric.updateValue(ts.timestamp)
           persistLatestProcessedSnapshotTimestamp(ts).map(_ => ts)
         }
     }
@@ -157,7 +180,7 @@ abstract class AcsSnapshotBulkStorage(
         Sink.ignore,
         automationConfig,
         backoffClock,
-        "ACS Snapshot Bulk Storage",
+        description,
         retryProvider,
         loggerFactory,
       )
